requirements.txt
flask==3.0.0
flask-cors==4.0.0
PyJWT==2.8.0
werkzeug==3.0.0
torch==2.1.0
librosa==0.10.1
numpy==1.24.3
python-dotenv==1.1.0
requests==2.31.0
